requirement.txt
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-decouple==2.1
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
multidict==6.0.4
openai==0.27.9
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
zope.interface==6.0
